move shared to helpers
