+++ test/canvas.test.ts
@@ -10,8 +10,8 @@ import * as users from '../src/libs/users';
 import * as calServer from './calendars/server';
 import config from './config';
 import { saveTestClass } from './helpers/class';
+import { buildRequest, requireLoggedIn, validateParameters } from './helpers/shared';
 import { generateJWT, saveTestUser, testUser } from './helpers/user';
-import { buildRequest, requireLoggedIn, validateParameters } from './shared';
 
 use(chaiSubset);
